Change the default behavior and update args accordingly
Usage now defaults to what were previously called -first and -from-env.
In my dotfiles usages, I was using these two options for all (but one!)
call to this utility, and it makes sense to make this the default. Most
usage will likely be for the PATH variable specifically, and script
usage makes it simpler to set as a string parameter instead of a pipe.

Bonus points: invocation of the command without a pipe will no longer
hang indefinitely waiting for standard input that will never come.
Reading from STDIN is now opt-in.

Also fixed behavior when given an empty existing PATH input. This is
probably rare in practice, but happens a lot in tests; I could
easily see a "start with nothing and build PATH up from scratch"
approach being useful in some cases.